# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CoachLink Platform API
  description: |
    API для платформы автоматизации тренировочного процесса
    и взаимодействия в системе «тренер — спортсмен».

    ## Аутентификация
    Все запросы (кроме /auth/*) требуют JWT access-token
    в заголовке `Authorization: Bearer <token>`.

    ## Роли
    - `coach` — тренер
    - `athlete` — спортсмен

    ## Формат дат
    - Даты в формате ISO 8601: `2026-03-28`
    - Timestamps в формате ISO 8601 с таймзоной: `2026-03-28T12:00:00Z`

    ## Пагинация
    Списочные эндпоинты поддерживают query-параметры `page` и `page_size`.
    Ответ содержит объект `pagination` с метаданными.
  version: 1.0.0
  contact:
    name: CoachLink Team
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidLogin:
                  summary: Невалидный логин
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: "login must contain only latin letters, digits and hyphens"
        '409':
          description: Логин уже занят
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                loginTaken:
                  summary: Логин уже существует
                  value:
                    error:
                      code: LOGIN_ALREADY_EXISTS
                      message: "User with login 'ivan123' already exists"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  value:
                    error:
                      code: INVALID_CREDENTIALS
                      message: "Invalid login or password"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление access-token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Невалидный или просроченный refresh-token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Выход (отзыв refresh-token)
      operationId: logout
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Успешный выход
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ----------------------------------------------------------
  # USERS
  # ----------------------------------------------------------
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: Список уведомлений текущего пользователя
      operationId: getNotifications
      security:
        - BearerAuth: []
      parameters:
        - name: is_read
          in: query
          required: false
          description: Фильтр по статусу прочтения
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Список уведомлений
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination, unread_count]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  unread_count:
                    type: integer
                    description: Общее количество непрочитанных уведомлений
                    example: 5
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/{notificationId}/read:
    put:
      tags: [Notifications]
      summary: Пометить уведомление прочитанным
      operationId: markNotificationRead
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: Уведомление помечено прочитанным
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/notifications/read-all:
    put:
      tags: [Notifications]
      summary: Пометить все уведомления прочитанными
      operationId: markAllNotificationsRead
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Все уведомления помечены прочитанными
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/device-token:
    post:
      tags: [Notifications]
      summary: Зарегистрировать FCM-токен устройства
      description: |
        Клиент отправляет FCM-токен для получения push-уведомлений.
        Если токен уже зарегистрирован — обновляет запись.
      operationId: registerDeviceToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceTokenRequest'
      responses:
        '200':
          description: Токен зарегистрирован
        '401':
          $ref: '#/components/responses/Unauthorized'

# ============================================================
# COMPONENTS
# ============================================================
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access-token, полученный при /auth/login или /auth/refresh

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
...
    NotificationId:
      name: notificationId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # ----------------------------------------------------------
  # Responses
  # ----------------------------------------------------------
  responses:
    Unauthorized:
      description: Не авторизован — невалидный или отсутствующий токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: "Invalid or expired access token"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: "Resource not found"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    RegisterRequest:
      type: object
      required: [login, email, password, full_name, role]
      properties:
        login:
          type: string
          pattern: '^[a-zA-Z0-9-]+$'
          minLength: 3
          maxLength: 50
          description: |
            Уникальный логин. Только латинские буквы, цифры и дефис.
          example: "ivan-petrov"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Минимум 8 символов
          example: "securePassword123"
        full_name:
          type: string
          minLength: 2
          maxLength: 255
          description: ФИО пользователя
          example: "Петров Иван Сергеевич"
        role:
          type: string
          enum: [coach, athlete]
          description: Роль пользователя в системе
          example: "athlete"
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          example: "ivan-petrov"
        password:
          type: string
          example: "securePassword123"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Refresh-token, полученный при входе
          example: "eyJhbGciOiJIUzI1NiIs..."
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Refresh-token для отзыва
    AuthResponse:
      type: object
      required: [access_token, refresh_token, expires_in, user]
      properties:
        access_token:
          type: string
          description: JWT access-token
          example: "eyJhbGciOiJIUzI1NiIs..."
        refresh_token:
          type: string
          description: Refresh-token для обновления access-token
          example: "dGhpcyBpcyBhIHJlZnJlc2g..."
        expires_in:
          type: integer
          description: Время жизни access-token в секундах
          example: 900
        user:
          $ref: '#/components/schemas/UserProfile'

    # === Users ===
    UserProfile:
      type: object
      required: [id, login, email, full_name, role, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        login:
          type: string
          example: "ivan-petrov"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        full_name:
          type: string
          example: "Петров Иван Сергеевич"
        role:
          type: string
          enum: [coach, athlete]
          example: "athlete"
        created_at:
          type: string
          format: date-time
          example: "2026-03-28T12:00:00Z"
...
    Notification:
      type: object
      required: [id, type, title, is_read, created_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - connection_request
            - connection_accepted
            - connection_rejected
            - training_assigned
            - training_deleted
            - report_submitted
            - group_added
            - group_removed
          description: Тип уведомления
          example: "training_assigned"
        title:
          type: string
          description: Заголовок уведомления
          example: "Новое задание от тренера"
        body:
          type: string
          nullable: true
          description: Текст уведомления
          example: 'Тренер Сидорова М.А. назначил вам тренировку "Развивающий кросс" на 05.04.2026'
        data:
          type: object
          nullable: true
          description: |
            Дополнительные данные для навигации в приложении.
            Содержит ID связанного ресурса для перехода при нажатии.
          properties:
            assignment_id:
              type: string
              format: uuid
            request_id:
              type: string
              format: uuid
            group_id:
              type: string
              format: uuid
          example:
            assignment_id: "550e8400-e29b-41d4-a716-446655440099"
        is_read:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
    RegisterDeviceTokenRequest:
      type: object
      required: [fcm_token]
      properties:
        fcm_token:
          type: string
          description: Firebase Cloud Messaging token
          example: "dGhpcyBpcyBhIGZjbSB0b2tlbg..."
        device_info:
          type: string
          nullable: true
          description: Информация об устройстве (модель, ОС)
          example: "iPhone 15 Pro, iOS 18.0"

    # === Common ===
    Pagination:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          description: Текущая страница
          example: 1
        page_size:
          type: integer
          description: Размер страницы
          example: 20
        total_items:
          type: integer
          description: Общее количество элементов
          example: 42
        total_pages:
          type: integer
          description: Общее количество страниц
          example: 3
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Машиночитаемый код ошибки.
                Коды: VALIDATION_ERROR, LOGIN_ALREADY_EXISTS, INVALID_CREDENTIALS,
                UNAUTHORIZED, FORBIDDEN, NOT_FOUND, ALREADY_HAS_COACH,
                REQUEST_ALREADY_EXISTS, ALREADY_IN_GROUP, REPORT_ALREADY_EXISTS,
                ASSIGNMENT_NOT_COMPLETED, INTERNAL_ERROR
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Человекочитаемое описание ошибки
              example: "login must contain only latin letters, digits and hyphens"
            details:
              type: array
              nullable: true
              description: Детали ошибок валидации (для каждого поля)
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "login"
                  message:
                    type: string
                    example: "must be at least 3 characters"